Extract fix seams into dedicated modules for rebase-friendly diffs
Per conflict hygiene: new code lives in new files so a future upstream
merge touches our modules plus a one-line import in theirs, instead of
rewriting regions inside upstream functions.

- tool-executor.ts: repairStringEncodedUnionMembers moves to
  util/repair-string-encoded-union-members.ts (call site unchanged).
- client.ts: mcpContentToToolResultOutputs moves to
  common/src/mcp/content-mapping.ts (call site unchanged).
- prompts.ts: serveInputSchema + ensureZodSchema move to
  tools/serve-input-schema.ts; prompts.ts drops the logger parameter
  added for the loud-fallback experiment and reverts
  ensureJsonSchemaCompatible to the upstream shape (452 lines, under
  the 500-line budget; ensureJsonSchemaCompatible remains upstream's
  silent-fallback version pending upstream buy-in).

Behavior unchanged: full suite 78/78.
