+++ packages/agent-runtime/src/tools/tool-executor.ts
@@ -10,6 +10,7 @@ import { formatValueForError } from '../util/format-value'
 import { codebuffToolHandlers } from './handlers/list'
 import { getMatchingSpawn } from './handlers/tool/spawn-agent-utils'
 import { getAgentTemplate } from '../templates/agent-registry'
+import { repairStringEncodedUnionMembers } from '../util/repair-string-encoded-union-members'
 import { resolveGravityIndexLink } from './gravity-index-cta'
 import { ensureZodSchema } from './prompts'
 
@@ -569,52 +570,6 @@ export async function executeToolCall<T extends ToolName>(
       )
     }
   })
-}
-
-/**
- * Repairs values the model string-encoded against its schema. When a
- * parameter's declared schema is a union containing an object variant, a
- * model may emit the object as a JSON-encoded string (a string is
- * unambiguously valid for the union, so nothing downstream fails). The
- * schema-guided decode below restores the object the model meant; plain
- * strings and params without an object variant are never touched, so
- * tools whose string parameters legitimately contain JSON (script
- * sources, file contents) are unaffected.
- */
-function repairStringEncodedUnionMembers(
-  parameters: Record<string, any>,
-  rawSchema: unknown,
-): void {
-  if (!rawSchema || typeof rawSchema !== 'object') return
-  const properties = (rawSchema as { properties?: Record<string, unknown> })
-    .properties
-  if (!properties) return
-  for (const [param, value] of Object.entries(parameters)) {
-    if (typeof value !== 'string') continue
-    const propSchema = properties[param]
-    if (!propSchema || typeof propSchema !== 'object') continue
-    const union =
-      (propSchema as { anyOf?: unknown[] }).anyOf ??
-      (propSchema as { oneOf?: unknown[] }).oneOf
-    if (!Array.isArray(union)) continue
-    const hasObjectVariant = union.some(
-      (variant) =>
-        variant &&
-        typeof variant === 'object' &&
-        (variant as { type?: unknown }).type === 'object',
-    )
-    if (!hasObjectVariant) continue
-    const trimmed = value.trim()
-    if (!trimmed.startsWith('{') && !trimmed.startsWith('[')) continue
-    try {
-      const decoded = JSON.parse(trimmed)
-      if (decoded && typeof decoded === 'object') {
-        parameters[param] = decoded
-      }
-    } catch {
-      // Not JSON after all - the string is a legitimate value.
-    }
-  }
 }
 
 export function parseRawCustomToolCall(params: {
